BE-641: Split the expression core and rewire the AST module tree
